converted the improper "id"s to "class"es
To be more exact the tags that were identified for applying styles
using id's were modified to be identified with classes as they are more
appropriate according to the HTML validator. The following message was
obtained from the validator before this commit,

"An "id" is a unique identifier. Each time this attribute is used in a
document it must have a different value. If you are using this attribute
as a hook for style sheets it may be more appropriate to use classes
(which group elements) than id (which are used to identify exactly one
element)."