Initialized the new project
The  project has been started anew without the
old history as it seemed the right choice because the
old history doesn't make much sense any more.

This project has been initialized with the files
taken created as a result of previous experience
from the old project. It stills exists in a lonely
place.