Don't elevate the main container for smaller screens
Elevating the main container for smaller screens seems
unneeded given the small screen size.

So, avoid elvating the container for saller screen sizes.

This obviates the need of different paddings for form-container
for different screen size as it's not elevated any more on small
screens where it didn't look nice.