form-2: distinguish buttons visually
It's better to distinguish the primary and secondary buttons
visually to express their importance more clearly.

This fixes the problem of using uncommunicative word for a button
to distinguish them as they communicative one was rhythmic with the
other one and would have confused users when the two buttons were
of the same colour.