form-1: update form sizing
As noted by @sirganeshv the main container was takin up
more width than it deserved. So reduce the width taken up
by the main container.

Credit goes to the designer of Google forms for giving an idea
about how to limit the form's width.

This update hasn't considered the space that would be taken up
by the status table.