Hygiene: HTML conformance fixes
The page is almost HTML conforming except for the 'action'
attribute of form having incorrect value which would be fixed.
Also, it does trigger for the invalid for meta element,

  Error: A meta element with an http-equiv attribute whose value
  is X-UA-Compatible must have a content attribute with the value
  IE=edge.

which I have chosen to ignore.

Surprisingly the HTML validator mentioned that 'button' cannot
have a type 'button' although the standard does allow it.
The exact error,

  Error: Bad value button for attribute type on element button.