Update header height
The percentage is really restrictive and dangerous for container sizes
as they get affected at the slightest variation of screen sizes like the
variation of screens size when a menu bar being added to the browser.

This is unwanted and overkill. So, avoid percentages and switch to pixels
until something better comes out.