major: make wepages dynamic
This adds the files required to make the project dynamic. These files
are taken from the remote repository[1] of a friend who takes care of the
backend of this project.

The 'html' counterparts of 'jsp' are kept just for reference. Only the
'jsp' files are dynamic, of course.

Note: Man it's way hard to find why you did something unless you've
recorded the history.

[1]: https://github.com/sirganeshv/SeminarRebuild